Made Create-Entry a component.
